requirements.txt
PySide2
certifi
chardet
idna
numpi
pillow
pyserial
pywin32
requests
scipy
shiboken2
urllib3